changed structure of response to have a current weather key
